requirements.txt
flask
Flask-SQLAlchemy
aichatbot